+++ src/components/filters/index.ts
@@ -1,0 +1,22 @@
+/**
+ * Filter components index
+ * Exports all specialized filter components
+ */
+
+export { default as BooleanFilter } from './BooleanFilter'
+export { default as MultiSelectFilter } from './MultiSelectFilter'
+export { default as RangeFilter } from './RangeFilter'
+export { default as FilterCategory } from './FilterCategory'
+
+export type {
+  FilterType,
+  FilterFacet,
+  FilterDefinition,
+  BaseFilterProps,
+  BooleanFilterProps,
+  MultiSelectFilterProps,
+  NumericFilterProps,
+  RangeFilterProps,
+  FilterCategoryProps,
+  Preset
+} from './types'